Change order for selecting department
When there is just a single entry in a select field that
is already selected the browser doesn't close the select field.
This causes an issue when a browser click is executed, because
instead of clicking the button it triggers the select field to close
instead.

So moving the select selection not to be last gets around this issue
we make it not the last action in the submit_form_func